+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix #3561: ensure okhttp resources are closed
 
 #### Improvements
+* Fix #3562: Kubernetes Mock Server improvements
 
 #### Dependency Upgrade
 * Fix #3562: Bump MockWebServer